# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectionSecretUpdate": {
        "properties": {
          "secret": {
            "maxLength": 16384,
            "minLength": 1,
            "title": "Secret",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "secret"
        ],
        "title": "ConnectionSecretUpdate",
        "type": "object"
      },
      "ConnectionStatusResponse": {
        "properties": {
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "last_test_status": {
            "anyOf": [
              {
                "enum": [
                  "success",
                  "failure"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Test Status"
          },
          "last_tested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Tested At"
          }
        },
        "required": [
          "configured",
          "last_test_status",
          "last_tested_at"
        ],
        "title": "ConnectionStatusResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReanalysisSettingsResponse": {
        "properties": {
          "creator_interval_days": {
            "title": "Creator Interval Days",
            "type": "integer"
          },
          "game_interval_days": {
            "title": "Game Interval Days",
            "type": "integer"
          }
        },
        "required": [
          "creator_interval_days",
          "game_interval_days"
        ],
        "title": "ReanalysisSettingsResponse",
        "type": "object"
      },
      "ReanalysisSettingsUpdate": {
        "properties": {
          "creator_interval_days": {
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Creator Interval Days",
            "type": "integer"
          },
          "game_interval_days": {
            "maximum": 90.0,
            "minimum": 1.0,
            "title": "Game Interval Days",
            "type": "integer"
          }
        },
        "required": [
          "creator_interval_days",
          "game_interval_days"
        ],
        "title": "ReanalysisSettingsUpdate",
        "type": "object"
      },
...
      "SessionResponse": {
        "properties": {
          "api_version": {
            "title": "Api Version",
            "type": "string"
          },
          "service_connections": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Service Connections",
            "type": "object"
          },
          "workspace_name": {
            "title": "Workspace Name",
            "type": "string"
          }
        },
        "required": [
          "workspace_name",
          "api_version",
          "service_connections"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Find Me Gamer API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/session": {
      "get": {
        "operationId": "validateSession",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Session",
        "tags": [
          "session"
        ]
      }
    },
    "/api/v1/settings/connections/{service}": {
      "get": {
        "operationId": "getConnectionStatus",
        "parameters": [
          {
            "in": "path",
            "name": "service",
            "required": true,
            "schema": {
              "title": "Service",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Connection",
        "tags": [
          "settings"
        ]
      },
      "post": {
        "operationId": "testConnection",
        "parameters": [
          {
            "in": "path",
            "name": "service",
            "required": true,
            "schema": {
              "title": "Service",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Test Connection",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "replaceConnectionSecret",
        "parameters": [
          {
            "in": "path",
            "name": "service",
            "required": true,
            "schema": {
              "title": "Service",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionSecretUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Replace Connection",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/reanalysis": {
      "get": {
        "operationId": "getReanalysisSettings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReanalysisSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Reanalysis",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "updateReanalysisSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReanalysisSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReanalysisSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Reanalysis",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}